papadin-ai/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.3.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
torch==2.1.1
pytesseract==0.3.10
opencv-python-headless==4.8.1
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0